+++ lib/simple_barcode_scanner.dart
@@ -1,9 +1,10 @@
 library simple_barcode_scanner;
 
-export 'package:simple_barcode_scanner/simple_barcode_scanner.dart';
 import 'package:flutter/material.dart';
 import 'package:simple_barcode_scanner/enum.dart';
 import 'package:simple_barcode_scanner/screens/shared.dart';
+
+export 'package:simple_barcode_scanner/simple_barcode_scanner.dart';
 
 class SimpleBarcodeScannerPage extends StatelessWidget {
   ///Barcode line color default set to #ff6666
@@ -18,14 +19,22 @@ class SimpleBarcodeScannerPage extends StatelessWidget {
   ///Enter enum scanType, It can be BARCODE, QR, DEFAULT
   final ScanType scanType;
 
-  ///All param only support for mobile and tab,
-  ///desktop and web is not yet support
+  ///AppBar Title
+  final String? appBarTitle;
+
+  ///center Title
+  final bool? centerTitle;
+
+  /// appBatTitle and centerTitle support in web and window only
+  /// Remaining field support in only mobile devices
   const SimpleBarcodeScannerPage({
     Key? key,
     this.lineColor = "#ff6666",
     this.cancelButtonText = "Cancel",
     this.isShowFlashIcon = false,
     this.scanType = ScanType.barcode,
+    this.appBarTitle,
+    this.centerTitle,
   }) : super(key: key);
 
   @override
@@ -35,6 +44,8 @@ class SimpleBarcodeScannerPage extends StatelessWidget {
       cancelButtonText: cancelButtonText,
       isShowFlashIcon: isShowFlashIcon,
       scanType: scanType,
+      appBarTitle: appBarTitle,
+      centerTitle: centerTitle,
       onScanned: (res) {
         Navigator.pop(context, res);
       },