+++ CHANGELOG.md
@@ -29,3 +29,8 @@
 ## 0.0.6
 * Document update
 * warning update
+
+
+## 0.0.7
+* Fixed performance issue with window [5](https://github.com/CodingWithTashi/simple_barcode_scanner/issues/5)
+* Added appBar title and centerTitle param [11](https://github.com/CodingWithTashi/simple_barcode_scanner/issues/11)